+++ behat.yml.dist
@@ -7,24 +7,21 @@ default:
             verbose: true
             paths: false
             snippets: false
+
     suites:
-            ui_homepage:
-                contexts:
-                    - sylius.behat.context.hook.doctrine_orm
+        ui_homepage:
+            contexts:
+                - sylius.behat.context.hook.doctrine_orm
+                - sylius.behat.context.transform.lexical
+                - sylius.behat.context.transform.product
+                - sylius.behat.context.transform.shared_storage
+                - sylius.behat.context.setup.channel
+                - sylius.behat.context.setup.product
+                - sylius.behat.context.setup.shop_security
+                - app.behat.context.ui.shop.homepage
 
-                    - sylius.behat.context.transform.lexical
-                    - sylius.behat.context.transform.product
-                    - sylius.behat.context.transform.shared_storage
-
-                    - sylius.behat.context.setup.channel
-                    - sylius.behat.context.setup.product
-                    - sylius.behat.context.setup.shop_security
-                    - sylius.behat.context.setup.homepage_context
-
-                    - sylius.behat.context.ui.shop.homepage
-
-                filters:
-                    tags: "@homepage && @ui"
+            filters:
+                tags: "@homepage && @ui"
 
     extensions:
         Lakion\Behat\MinkDebugExtension:
@@ -59,6 +56,10 @@ default:
             show_auto: false
 
         FriendsOfBehat\SymfonyExtension: ~
+
+        FriendsOfBehat\ExcludeSpecificationsExtension:
+            features:
+                - "vendor/sylius/sylius/features/homepage/viewing_latest_products.feature"
 
         FriendsOfBehat\VariadicExtension: ~
