+++ config/packages/_sylius.yaml
@@ -146,7 +146,7 @@ sylius_product:
         product:
             classes:
                 model: App\Entity\Product\Product
-                repository: App\Repository\ProductRepository
+                repository: App\Doctrine\ORM\ProductRepository
             translation:
                 classes:
                     model: App\Entity\Product\ProductTranslation
