+++ config/packages/test/behat.xml
@@ -3,10 +3,9 @@
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
         <defaults public="true" />
-        <service id="sylius.behat.context.setup.homepage_context" class="App\Tests\contexts\shop\HomepageContext">
+        <service id="app.behat.context.ui.shop.homepage" class="App\Tests\Behat\Context\Shop\HomepageContext">
             <argument type="service" id="sylius.behat.page.shop.homepage" />
         </service>
-        <service id="sylius.behat.page.shop.homepage" class="App\Tests\Behat\Page\shop\HomePage" parent="sylius.behat.page.shop.home" public="false">
-        </service>
+        <service id="sylius.behat.page.shop.homepage" class="App\Tests\Behat\Page\Shop\HomePage" parent="sylius.behat.page.shop.home" public="true" />
     </services>
 </container>
