+++ config/services_test.yaml
@@ -1,2 +1,3 @@
 imports:
     - { resource: "../vendor/sylius/sylius/src/Sylius/Behat/Resources/config/services.xml" }
+    - { resource: "packages/test/behat.xml" }
