+++ templates/bundles/SyliusShopBundle/Homepage/index.html.twig
@@ -4,5 +4,5 @@
     <h2 class="ui horizontal section divider header">
         Clearance sale
     </h2>
-    {{ render(url('app_admin_custom_product_index')) }}
+    {{ render(url('app_shop_partial_product_index_by_on_hand')) }}
 {% endblock %}
