+++ tests/Behat/Page/Shop/HomePage.php
@@ -7,9 +7,9 @@ namespace App\Tests\Behat\Page\Shop;
 use Behat\Mink\Element\NodeElement;
 use Sylius\Behat\Page\Shop\HomePage as BaseHomePage;
 
-class Homepage extends BaseHomePage
+class HomePage extends BaseHomePage
 {
-    public function getProductsNameFromHomepage(): array
+    public function getProductsNamesFromHomepage(): array
     {
         return array_map(
             function (NodeElement $element) {
